+++ packages/devtools_app/lib/src/http/http_request_data.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import '../timeline/flutter/timeline_model.dart';
+import '../trace_event.dart';
 import '../utils.dart';
 
 import 'http.dart';
