+++ packages/devtools_app/lib/src/profiler/cpu_profile_model.dart
@@ -6,6 +6,7 @@ import 'dart:convert';
 
 import 'package:meta/meta.dart';
 
+import '../trace_event.dart';
 import '../trees.dart';
 import '../utils.dart';
 
@@ -50,13 +51,9 @@ class CpuProfileData {
   ) {
     // Each trace event in [subTraceEvents] will have the leaf stack frame id
     // for a cpu sample within [subTimeRange].
-    // TODO(kenz): use TraceEvent.timestampKey instead of 'ts' once we delete
-    // the html app. Attempting to import `timeline_model.dart` was causing
-    // dependency issues when trying to use this file in both the html and
-    // Flutter versions of DevTools.
     final subTraceEvents = superProfile.stackTraceEvents
-        .where((trace) =>
-            subTimeRange.contains(Duration(microseconds: trace['ts'])))
+        .where((trace) => subTimeRange
+            .contains(Duration(microseconds: trace[TraceEvent.timestampKey])))
         .toList();
 
     // Use a SplayTreeMap so that map iteration will be in sorted key order.
