+++ packages/devtools_app/lib/src/timeline/html_timeline_processor.dart
@@ -8,6 +8,7 @@ import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
 
 import '../config_specific/logger/logger.dart';
+import '../trace_event.dart';
 import '../ui/fake_flutter/fake_flutter.dart';
 import '../utils.dart';
 //import 'simple_trace_example.dart';
