+++ packages/devtools_app/test/flutter/timeline_processor_test.dart
@@ -5,6 +5,7 @@
 import 'package:devtools_app/src/timeline/flutter/timeline_controller.dart';
 import 'package:devtools_app/src/timeline/flutter/timeline_model.dart';
 import 'package:devtools_app/src/timeline/flutter/timeline_processor.dart';
+import 'package:devtools_app/src/trace_event.dart';
 import 'package:devtools_app/src/utils.dart';
 import 'package:devtools_testing/support/flutter/test_utils.dart';
 import 'package:devtools_testing/support/flutter/timeline_test_data.dart';
