+++ packages/devtools_app/test/timeline_model_test.dart
@@ -6,6 +6,7 @@ import 'dart:convert';
 import 'package:devtools_app/src/profiler/cpu_profile_model.dart';
 import 'package:devtools_app/src/timeline/html_timeline_controller.dart';
 import 'package:devtools_app/src/timeline/html_timeline_model.dart';
+import 'package:devtools_app/src/trace_event.dart';
 import 'package:devtools_app/src/utils.dart';
 import 'package:devtools_testing/support/cpu_profile_test_data.dart';
 import 'package:devtools_testing/support/test_utils.dart';
