+++ packages/devtools_app/test/timeline_processor_test.dart
@@ -5,6 +5,7 @@
 import 'package:devtools_app/src/timeline/html_timeline_controller.dart';
 import 'package:devtools_app/src/timeline/html_timeline_model.dart';
 import 'package:devtools_app/src/timeline/html_timeline_processor.dart';
+import 'package:devtools_app/src/trace_event.dart';
 import 'package:devtools_app/src/utils.dart';
 import 'package:devtools_testing/support/test_utils.dart';
 import 'package:devtools_testing/support/timeline_test_data.dart';
