+++ packages/devtools_testing/lib/support/flutter/timeline_test_data.dart
@@ -8,6 +8,7 @@ import 'dart:convert';
 import 'package:devtools_app/src/timeline/flutter/timeline_controller.dart'
     show timelineScreenId;
 import 'package:devtools_app/src/timeline/flutter/timeline_model.dart';
+import 'package:devtools_app/src/trace_event.dart';
 
 import '../cpu_profile_test_data.dart';
 import 'test_utils.dart';
