+++ packages/devtools_testing/lib/support/test_utils.dart
@@ -7,6 +7,7 @@
 import 'dart:convert';
 
 import 'package:devtools_app/src/timeline/html_timeline_model.dart';
+import 'package:devtools_app/src/trace_event.dart';
 
 SyncTimelineEvent testSyncTimelineEvent(TraceEventWrapper eventWrapper) =>
     SyncTimelineEvent(eventWrapper);
