+++ packages/devtools_testing/lib/support/timeline_test_data.dart
@@ -8,6 +8,7 @@ import 'dart:convert';
 import 'package:devtools_app/src/timeline/html_timeline_controller.dart'
     show timelineScreenId, TimelineMode;
 import 'package:devtools_app/src/timeline/html_timeline_model.dart';
+import 'package:devtools_app/src/trace_event.dart';
 
 import 'cpu_profile_test_data.dart';
 import 'test_utils.dart';
